# Inventory pagination query performance improvement